warning: use of a signed integer operand with a binary bitwise operator

+++ src/search16.cc
@@ -481,7 +481,7 @@ auto backtrack_16(char * qseq,
       const uint64_t d
         = dirbuffer[(offset
                      + longestdbsequence * 4 * static_cast<uint64_t>(j / 4)
-                     + static_cast<uint64_t>(4 * i + (j & 3U)))
+                     + static_cast<uint64_t>(4 * i + (static_cast<uint64_t>(j) & 3U)))
                     % dirbuffersize];
 
       if ((operation == 'I') && ((d & maskextleft) == 0U))
